The interpreter can be interrupted with additional code to run
__new__ and __init__ spurred doing this. It looks like the Python interpreter
does something like this when it needs the result of Python calls to proceed to
do work like create an object. I sketched what that might look like, but it's
not enabled. I instead just updated the interpreter to handle this without
previous code failing.

The wait test did fail, however, since this kind of screwed up internal state
it was tracking. However, that whole mechanism was hopeless crude (just
maintaining the current cursor and frames) and will need to be done better to
create green threading.