We can now create classes (that don't have constructors)
Had to bring the cursor properties back into the interpeter. We cannot yet call
a user-defined constructor since it doesn't know to implicitly pass in self.

There are limitations with the interactions between built-ins and the
interpreter. Return calls now return from the interpreter Run() call instead of
keeping the loop going. So we have a kind-of iterative model for running Python
calls along with a recursiving method. It's only through luck right now that
that FunctionTests still works.