Started implementing tasklets using IEnumerables.
The interpreter now is meant to be run within the control of a scheduler.

Separated runtime context from interpreter state. The interpreter is
mostly stateless now. What it is running at any moment is stored in a
FrameContext, which contains the stack of frames. This means we no longer
fidget with the interpreter when changing what we're running.

Unit tests work again with this contextless interpreter wrapped inside a
scheduler.

Updating documentation to mention tasklet implementation