Separated runtime context from interpreter state
The interpreter is mostly stateless now. What it is running at any moment is
stored in a FrameContext, which contains the stack of frames. This means we no
longer fidget with the interpreter when changing what we're running.

The demo test does run but I haven't vetted the unit tests yet.