Came up with a rough common abstraction for calling code
This unites:
* CodeObject
* WrappedCodeObject
* PyTypeObject (PyClass) that is called to create an instance

The interpreter, frame context, and arguments are passed in. WrappedCodeObject
doesn't use the interpreter, but will need the frame context to properly
construct exception stack traces (when we get there).

A lot of old code is commented-out, but still present, and the new code isn't
documented yet. I'm still figuring it out.