Starting implementation of raise statement with a PyException type
Came up with a rough common abstraction for calling code

This unites:
* CodeObject
* WrappedCodeObject
* PyTypeObject (PyClass) that is called to create an instance

The interpreter, frame context, and arguments are passed in. WrappedCodeObject
doesn't use the interpreter, but will need the frame context to properly
construct exception stack traces (when we get there).

Naive RAISE_VARARGS implementation raises exceptions now

RaiseException test passes. However, this is a naive implementation that just
throws the exception out. It doesn't internalize it for catching in the script
(yet).

AttributeError is a PythonException subclass.

Refactored DataTypes into individual files in subdirectories. It had finally
become too unwieldy to work with them in one file.