We can now handle catch-all exception clauses.
TryExceptBlank passes now.

Implemented SETUP_EXCEPT and basic reactions to exception blocks. We hold
the current exception in the frame context and inspect that for an active
exception. I believe Python does something similar. If there's an exception, we
start unrolling our blocks. If we don't have any to unroll then we'll return
without doing anything with the exception.