Starting to handle a raised exception instead of immediately throwing it
We don't go to the except blocks yet. Sooon. This just introduces some of the
handling. The original RaiseException test still works with this handling.

We hold the current exception in the frame context and inspect that for an
active exception. I believe Python does something similar. If there's an
exception, we start unrolling our blocks. If we don't have any to unroll then
we'll return without doing anything with the exception.