Except clauses can now qualify by type of exception
TryExceptTyped passes. It wasn't checked in aggressive detail so it may be a
false pass.

Added the ExceptionMatch comparison operation. It makes sure that we are
comparing to an exception and then that their classes match.

END_FINALLY drafted. It was generated by the Python interpreter when this I
tested how Python generated this code. It doesn't appear too necessary but I
went along with it.