Start of a byte code debugger that steps through the code and dumps state
It takes a script as a command-line parameter, builds it, and then allows the
user to step through it and see how the interpreter behaves. I decided to do
this now because the exception handling logic was getting complicated and it was
hard to follow in Visual Studio's debugger directly breaking into the C# code.

Some augmentations was made to the interpreter and scheduler to accommodate
stepping and dumping.

The interpreter's UX is using LibREPL for a shell. This is courtesy of:
https://github.com/piksel/LibREPL

The library isn't something like a NuGet package or anything, so I had to
drop the code in wholesale. MIT license.

There's some issues with requiring multiple steps that seem inconsequential
before being prompted that the scheduler is done. This kind of thing might be
real issues, so I want to investigate them separately.

The error listener was encapsulated in one file because it is now used in three
different places.