Was not breaking out of interpreter block after function call. Fixed.
Noticed this while step debugging. After creating the exception class, it kept
going to raise it.

There are still problems with the test, and the interpreter is raising the
exception in step debugging instead of letting me walk through the proper,
internal handling. This is just a tweak along the way.