try-finally test now passes
Had to monkey with the unit test bench a little bit to make the frame context
an output variable. It would be null normally if run with an (expected) escaped
exception. As an output variable, it can be set before the exception escapes
so that the context can be inspected for assertions.

The interpreter is doing some quirky handling for END_FINALLY that needs to
be researched.

Fixups for the try block in the visitor still had some bugs and had to be
tweaked.