Adding a bytecode fixup container
This is a helper object to calculate the jump targets once they've been
determined during code generation. I was getting annoyed with all the integer
shenanigans with jump offset fixups in bytecode generation, so I'm testing out a
fixup container to do the work.

Started using the jump opcode fixer in try-catch

Consolidated context text checks with a hasText() helper

Not much of a reduction in code but it is less ugly! =D