Started (grossly) injecting Python *args to .net params
__build_class__ needed params to support the *args field that it has in
CPython. That field is used to denote subclasses. Since we're starting to
subclass, we have to start doing something else similar.

Encapsulated param conversion in WrappedCodeObject. That's the better place to
be making these decisions than one spot in the interpreter.

Some of the magic number shenanigans have been improved. Some of that comes
from the encapsulation and seeing what's going on the implicit treatment of
FrameContext.