Knocked around a ton of drywall to plumb FrameContext to super()
Added a TODO to inject FrameContext based on MethodInfo because the
shenanigans in WrappedCodeObject is wrong. Getting the context on construction
works for local calls, but not for builtins. The context is being passed during
the call and we just need to know to carry it.

I will deal with that once I know if this is even the right thing to be doing
for super().