Fixed up VisitAtom_expr to chain commands
We needed to make it look up trailers because of our super call:
super().__init__()

This was a mess to try to incorporate in the existing code so it was
overhauled. SubclassSuperconstructor still fails now but it looks to be an
issue with sprucing up PyClass to be a PyObject (I think).