Made PyClass a PyObject
This made SubclassSuperconstructor pass--at the expense of a few other OOP
unit tests. It turns out I have not been managing self correctly for a lot of my
code. I should not be putting self on the stack when generating method calls.
I should be creating bound methods instead. I will dealing with that now...