Introducing PyMethod to curry self to object functions
This is similar to how CPython does it. There is a PyMethod type that carries
self forward. How exactly it is done likely really differs from CPython. When
the object dictionary is created for an instance, the class code objects get
wrapped into PyMethods.

This got the rest of the object tests working again!