Drafted type(), isinstance(), and issubclass()
I think I need some of this stuff for determine children of Exceptions, so I
decided to take the intermediate step of introducing these built-ins.

I haven't delved into the obscura of how these are implemented in CPython, so
my implementations are without a doubt incomplete, but they do the basic
things in the interpreter.