The interpreter is starting to recognize Exception subclasses
This started working with tweaks to the ExceptionMatch comparison opcode. We
now match first to a generic PyClass and then use issubclass to make sure we
are comparing to an exception in general.

This caused some regressions for the built-in Exception type because issubclass
would not consider Exception to be a subclass of Exception. I mean, in English
it isn't, but it is logically True in Python. So issubclass was extended so
that equal types are regarded as subclasses.

TryExceptAliasUseValue now passes.
The similar TryExceptTwoExceptions does not yet; the problem appears to be in
byte code generation and will require different effort to enable.