Sketches for transition to async-await
Using IEnumerable coroutines requires copying and pasting a pretty regular
schedule block for every call that does it and every intermediary make
calls through a callable. We now have a trial project demonstrating how
we could use async-await instead with a secondary scheduler. We'll be
switching to that with the trial project as a proof-of-concept.

This also opens us up to be able to serialize our place in such scripts
and resume the commands.