Set up WAIT to use async-await, but code is still dark
We will have to mess with the return signature of the interpreter's Run method
and that will cause a cascade of refactors to remove the IEnumerables. I
decided to pause here first.

YieldTick was taken from SyncContextTest nearly wholesale. SetYielded was also
taken from it; I finally see where it was actually needed... kind of.