Switched from IEnumerables to async-await for concurrency
This is a complete implementation. Previous tests and tools should work. The
scheduler works using some clever management of tasks. We use an
ISubscheduledContinuation to notify the scheduler when the current call is going
to block. This sets it aside in the schedule to run something else instead. This
will allow (wrappers for) external agents to block scripts until their results
walk in. In particular, this gives me the means I wanted for NPC dialogs in a
game project to block scripts without hanging everything.

The fickle nature of exception handling with tasks required some tweaks. We
added a second slot into a frame context to handle escaped, general-purpose .NET
exceptions. They will get rethrown by the scheduler.

The async tasks will raise AggregateExceptions and we'll have to unpack them.
We also have to use tasks all the way up and out of RunUntilDone so we can
Wait on them and get these exceptions. If we don't wait on them, they won't
get thrown all the way up.

This is the sum of work performend on the async_await_impl branch.