The testbench properly receives escaped exceptions again
The async tasks will raise AggregateExceptions and we'll have to unpack them.
We also have to use tasks all the way up and out of RunUntilDone so we can
Wait on them and get these exceptions. If we don't wait on them, they won't
get thrown all the way up.