The debugging tool now steps properly again (with exception handling)
First, when in step mode, the active task will YieldTick after every iteration
of the interpreter loop.

Second, it was throwing a NotImplementedException that was just falling into
the ether before this was implemented. Not good. We added a second slot into a
frame context to handle escaped, general-purpose .NET exceptions. They will
get rethrown by the scheduler.