builtins__build_class returning Task<PyClass> is The Devil
This wasn't getting recognized as Task<object> and would not get unpacked.
Rather, the task would go on the data stack as the output. This blew up a huge
pile of unit tests that seemed to work before. It looks like this screwup
wasn't rebuilt so I was goaded into thinking my code base was healthier! :(