Filled out PyTuple, which dragged in proper PyString management
The PyTuple test failed since it expected PyObjects but I still was mixing in
regular strings. Making that take in a PyString caused a cascade of problems--
particularly with exception management.

PyString has an implicit operator to convert to string, but that doesn't work
for invoking MethodInfos, so I added overloaded functions to take PyStrings.