Added more whitespace-related grammar from original definition
There's a test to check how this work with a particularly bad ListReadWrite
declaration. I found out that's why the newline logic is written how it is.
That is why the newline stuff doesn't work in the REPL. The test passes
as-is, but if I loosen up that logic, then the REPL will start to work... but
the test will fail.