Implemented __repr__ for basic types
We now use __repr__ in the REPL to print the result. The __str__ dunder is
also implemented with a default, base implementation.

Calling __repr__ adding some async-await to the REPL demo Windows form
application.

I had to engage in some shenanigans to get the PyClass polymorphism for
__repr__ to work correctly. The reflection to get all the methods that have the
ClassMethod decorator got much more sophisticated.
1. I had to pass in binding flags to get parent functions.
2. I then had to make sure the final dictionary had the lowest subclass
   implementation as the one it'll actually call.