Refactored most of the ISubscheduledContinuation to require just a scheduler
They required an interpreter before, from which it extract the scheduler. The
interpreter was a concrete type, not an interface. So the scheduler was made
an interface too. The interpreter and scheduler interfaces were moved into
the language implementation because some things there needed awareness of
them, but I didn't want to put the actual class code in there.

This came up when I started trying to use a void sentinel continuation in the
Windows form application in the sleep wrapper. That code is unimplemented yet.
This commit just gets the rest of the source base positioned so I could make
that code compile.