Temporary fix for scheduler hangs from REPL demo sleep command
It looks like the Task.Delay calls happen in another thread or at least
out-of-order. This screws up using the last scheduled record to tell which
task we are getting a completion from. I am not sure if that approach was even
robust in the first place, so we may have to examine this in more detail.