FutureAwaiter results are now unpacked for the data stack by the interpreter
I added IGetsFutureAwaiterResult because I didn't have a general-purpose call
I could use to extract the future's result; generics don't take well to
polymorphism like that. Since on the sum I only needed to get a handle to the
object, I created that interface and forced FutureAwaiter to implement it. The
interpreter just uses that to get the value it needs.