REPL demo dialog now tries to wait for a choice; currently hangs
The REPL hangs waiting for the interpreter to finish, so we get a deadlock
because this blocks the GUI from updating with the choices and acknowledge the
selection. We need a way to interact with the schedule to leave the scheduler
loop if all tasks are blocked, and then a way to come back in afterwards when
it unblocks.