sleep() and dialog() now work in REPL demo
Repl can now be triggered to run from embeddables. This is used to kick the
interpreter into completing when a blocking routine unblocks. Without this,
the interpreter never gets a chance to finish the unblocked code.

Got rid of some thread forks in awaiters.

The problem appears to be with ConfigureAwait(). I guess I was setting false in
the demo because the awaits were locking up the GUI. However, this was
causing the resumptions to happen on other threads. Hilarity ensued...
generally resulting in a hang from mismatched state.

Adding TODO for introducing a thread ID guard to scheduler to help point out
in the future when stuff like this is happening.