Got rid of some thread forks in awaiters
This enabled both sleep() and dialog() in the REPL demo! We are freezing here
to capture this success.

The problem appears to be with ConfigureAwait. I guess I was setting false in
the demo because the awaits were locking up the GUI. However, this was
causing the resumptions to happen on other threads. Hilarity ensued...
generally resulting in a hang from mismatched state.