if-else blocks properly set up forward jumps now
Else clauses were not getting run at all.

This was fixed with a bit of pecking and could still have problems. It just
seems to be okay now. What makes this funnier is the SingleLayerIfElifElse
test had a wrong assertion in it that created a false pass. This was exposed
after making the fix. So I had to also change that assertion.