Added helper to use PyBool true/false static instances
The get_blip_wrapper in the REPL was creating new PyBools instead of using the
existing True/False instances. This would cause the check in UNARY_NOT to
fail. I added this helper to make it easier to do the right thing. Ultimately,
we should resort to a factory to create the objects correctly.