Starting to convert between Python and .NET types for function calls.
This is the result of a lot of work in the mix_py_and_csharp_objs branch that
has matured to the point of coming to master. This was squashed from the
commit: 522ce107ee59ed3d37f08329393e7d8f635ba7bd

The injector has been expanded to convert a suite of Python objects to
their .NET equivalents. This means the injector always runs. It also
handles params syntax. It does not otherwise work on lists or other types
yet. The simple types are accommodated in a basic converter helper.

An object resolver was added to help poke .NET objects for how to embed them
into Python. This deals with methods, fields, properties, and other junk that
has to be reflected.

WrappedCodeObject was moved to its own file because it is become more
sophisticated.

Added NotImplemented.