WIP saving sketchwork to call a WrappedCodeObject from an event
Based on what I saw in IronPython, I'll have to use a code generator to create
the binding that'll invoke Call() in my callables. There isn't a way to use a
lambda. This bit with a DynamicMethod seems to be the way.

I am just pausing it here in a branch while I look at some other code.