Reproduced situation of calling C# methods more conventionally
The self argument does not get carried to the WrappedCodeObject. We need to
figure out how to include it. It looks like it will have to get paired when
doing LOAD_ATTR. The first way we do this will probably be too ham-fisted for
final use, but it'll show if we have a handle on the problem.