New object instances once again properly create proper wrapped PyMethods
A bad refactor had it checking for CodeObject and missing WrappedCodeObjects.
This is why InvokingWrappedMethod failed. It had to test for IPyCallable
instead. Fixed.

Various other code had to be adjusted to remove self pointers they were
inserting to get around this problem. It appears this has been a problem for
awhile and we just coded around it without realizing!

To make it extremely certain that all objects had their methods properly bound,
I had to write a Create() helper to make all the instances. It's taboo to
directly create Python objects due to this! I think we'll want to devise a
better way to handle this shortly.