First working demonstration of event subscribe/unsubscribe
Note that this isn't a full-production implementation yet, but it does work
for at least that limited EventDotNet scenario.

The += and -= operators now handle .NET event subscribe/unsubscribe.

This was a nontrivial undertaking. We had to present an IPyCallable to a .NET
event. The event expected a specific signature with parameters ordered in a
typical .NET method, but the callable expected an object[] for it's Call()
method.

I had experimented with a DynamicMethod. This would have required generating
the code using MSIL. I was starting to see some success in this with general
prototyping. However, some discussions online got me going towards "abusing
generic methods."

The CallableDelegateProxy contains the requisite abused generic methods. These
are realized as the method signature I need when I need to bind to an event.
The proxy was necessary because I needed to include the IPyCallable in the
invocation. So I present a method with func(A, B...) that internally has the
callable, packs the args into an object[] and invokes Call() with that array
as the arguments.

The interpreter then had to be updated to support subscribe/unsubscribe. I
added EventInstance to the ObjectResolver to carry all the proper context for
subscribing and unsubscribing. Unsubscribing is particularly difficult because
I need to perfectly match the delegate I subscribe. I include the event's
delegate to help keep this up. The object resolver creates EventInstance when
it finds it's referencing an event in a .NET object. That's what winds up on
the interpreter stack.