PyMethods are now dynamically-generated, which forced adding PySuper too
The Python data model states that PyMethods are generated each time they're
needed. We were making them up front. That's theoretically more efficient,
but we were loading up __dict__ with all of these things. It also wasn't
quote right for inheritance. So we fixed it.

This opened up a GIANT CAN OF WORMS with super() and inheritance chains.
I found out that super() returns an object of a special type (PySuper_Type)
that I mimick. However, it's a shift little chameleon I don't yet understand.
I am currently using a hack with __getattribute__ to have it pass along
superclass methods as it's expected to do on lookups. This isn't quite how
CPython does it but I can't deduce what it does exactly.

I will have to figure out what that is and resolve it. The good news is that
the current hack is passing tests. For all the hackery, it's now generating
PyMethods on-demand when doing attribute lookups. It's just that PySuper is
an open book that hasn't been finished.