Hacked together some PySuper attribute overrides to properly resolve self, class
PySuper is doing some gross stuff right now, but I can't for the life of me
figure out how CPython takes care of stuff with super. It looks like it passes
along SOME attributes but not all. __dict__ is self.__dict__. So I currently
override __getattribute__ to rejigger some plumbing, including PyMethod
hackery. CPython seems to do something like __getattr__, but that is not
defined. It's doing ... something else.

I will have to figure out what that is and resolve it. The good news is that
the current hack is passing tests. For all the hackery, it's now generating
PyMethods on-demand when doing attribute lookups. It's just that PySuper is
an open book that hasn't been finished.