Classes are now created by using __call__ on the class
Previously, the constructor was wrapped code. We're trying to put all our
classes directly on the stack while still being able to invoking the
constructors. The interpreter uses <class>.__call__ for this, so we
implemented that as well.

PyMethod creation initially got in the way because we'd create those for every
function returned from the stack. That's not the correct thing to do. For now,
we skip wrapping __init__ but there's probably more we have to do foreach
staticmethod and stuff like that.