Initial, naive implementation of module imports.
Created a PyModule type that is pretty much just a PyObject whose
dictionary is stuff we import. We can feed a dictonary of these into the
interpreter and it can import from them. This is a far cry from a complete,
robust import system. We can't actually import other source files in a
path. We don't have a .NET import mechanism. Our unit tests mostly show
that the operations don't destroy themselves but we don't prove we can use
the imported stuff. However, it's a proof-of-concept initial step.

Implemented:
IMPORT_NAME
IMPORT_FROM
IMPORT_STAR

A common code generator was written to handle the three cases together
based on the slightly-different-but-similar grammatical parsing rules.

A good next step would be to create module specs for being able to define
how to import built-in stuff versus other Python code in some file system
vs whatever else we want. .NET imports would probably be handled with
something like a clr module like Python.NET/IronPython. This is TBD.