Implemented the most basic of dumb module imports
No module specs or much of anything. We just have a lookup of PyModules and
IMPORT_NAME is able to resolve them. Added more plumbing to the test bench to
support this. I now declare the test bench to be too cumbersome and all those
parameters need to move to a builder.