Injected module loader no longer double-searches when loading the module
I found out from PEP-451 that there's a miscellaneous loader_state property
where I can stuff specifics, so I put the injected module there. This will
be handed off when Load is invoked.

It's not supposed to be a (premature) optimization. I mostly wanted to just
add loader_state and take advantage of it.