Working prototype of a file-based module finder/loader
FileBasedModuleLoader.cs contains the finder and the loader.

ISpecLoader now takes the interpreter and a context; it is also now async.
The interpreter might yield while running the imported code. This also had
some effect on IMPORT_NAME and unit tests.

Added a unit test for file importing. This one even asserts that the module
has the right stuff in it!