Setting up more object extraction tests based on types
These came up when we started importing .NET types. We're trying to look at
System.Environment.MachineName. System.Environment is what's imported, and
it's a type. MachineName is a static field. We need to buff up the
ObjectResolver to handle this. It may have to do this while not entirely
Ignoring Type as a thing we might resolve too...