The object resolve can resolve from types now and not just instances
This passes the ImportSystemEnvironmentInCode test. The field we were trying
to resolve was static and System.Environment was a type. The ObjectResolver
would puke on it. It was changed to test if it was looking up a type already.
If it was, it would just work with the type. Otherwise, it would turn it into
a type like before.

A bunch of rawObject.GetType() calls were refactored to use the objType field
due to this extra resolution. It would arguably also be more performant than
calling GetType() each time we wanted the type, although the compiler probably
already took care of that.