Module loaders now might load a direct asset instead of a module
I got a hell of a corner case here with the CLR module importing a class
directly from a null namespace. You just use the import statement, not
import-from so it would get mixed up.

I'm asking online if it's even a thing for a module loader to do this. If not,
then what is Python.NET actually doing?

The loader is extremely inefficient with this, by the way.

The Unity test involving this doesn't completely work yet, by the way, but this
moved it forward.