Simplified the scheduler's completion checks
(actually, it's more like "wait how did this work in the first place?!" The
Pepe Sylvia logic that would put tasks back into the active queue wasn't
triggering even though it should have. With the logic simplified, it did, but
according to the logic, we'd want to mark the task as completed instead. So
the whole block was turned inside-out.

Scheduling new tasks inside code still doesn't work, by the way.