Apparently scheduled functions don't end up running the parent context
...this test was supposed to show a bug, but it didn't! The subcontexts we're
creating don't have anything in them to show what level of the stack is where
they should end. So when the function finishes, I expect the interpreter to run
the parent context. This would cause stuff to be run multiple times when it
should only be run once. Yet... this doesn't happen? Eh?

(I did have a conversion bug though.)