Refactored all types to use a common "InternalValue" name for sake of dynamics
There are a few operations that look at the Py* types as dynamic type. A
problem came up with using '+' for string concatenation here because the helper
assumed a numeric type (PyFloat, PyInteger) that has the .number field. It
failed with a string because it's internal representation was not using
.number. Whoops.

Now they all use "InternalValue." It's vague but these dynamic helpers can all
munge on it. This made it possible to easily mix .NET types with PyStrings
in printing/logging statements in embedded scripts.