Implemented the return statement
How the hell did I get this far without this? The interpreter would just scrape
whatever was on top of the data stack to return. Worked fine when I had to
return something. Didn't do squat when I, uh, had nothing to return. Cool.

A lot of bad tests started passing when I implemented this! =D