Fixed a lot of local/localname mismanagement
Copies of locals would get declared, many with null values. This came up when
the REPL would crashing when rewriting an existing function. Most locals were
added through a basic append instead of looking for existing entries to index.

To be honest, I don't have complete confidence, but it does pass all the
tests so far.