Scheduler will call received IPyCallables even if they aren't CodeObjects
The user might not know that they're trying to schedule something that isn't
interpreter code, but rather a WrappedCodeObject. Given the dynamic nature of
the language, this may be non-trivial. So we just call them when we get them.

Eventually, we'll need to create task records for them. That has been added as
a TODO.

This concluded the bulk of open TODO items so the TODO was greatly cleaned up.