Added test to show we can't use imported System.Collections.Generic stuff
This was found during Unity integration testing. It turns out the names that
are used include a suffix to show the number of generic arguments they have.
This doesn't resolve. We started to support importing the stuff but we can't
invoke their constructors yet--probably since they have the screwy name too.
That's the work-in-progress.

The solution of just snipping off the suffix is not really robust either and I
added a TODO to make that more robust.