Support for .NET generic constructors
This was an oversight that was exposed in Unity integration testing. I wanted
to use List<T> as a workaround for another problem--since fixes--and I
discovered it was outright impossible to import it, let along construct it.
Importing was handled in the previous commit, and this handles construction.

WrappedCodeObject continues to get more complicated. I'm inclined to wait even
further to simplify this code because I keep finding gotchas from integration.