Fixed an issue with selecting the wrong .NET method when binding
It wasn't really comparing parameter sizes correctly. This came up with
Object.Destroy() in Unity. It was selecting the "first" one that has two
parameters, but we wanted the one with only one field.

Fixing this caused an issue with properly finding .NET methods with optional
fields that also had to be fixed.