We can now try (and fail) to import heapq in the demo application
It's available to be imported, but it's Error Vomit City, so now we get to
fix and harden all kinds of crap. This is somewhat expected though. This is
the first time I'm trying to parse and run somebody else's non-trivial code.

Also, the demo application is using my hardcoded path to
PythonStandardLibrary, and we'll eventually have to correct that.