Basic support for variable arguments (*args) in functions
Added tests to support variable arguments and the unpacking operator. The
unpacking operator tests are still stubbed out. Some TODO items were left to
deal with them. It may not be necessary to understand CALL_FUNCTION_EX to do
them, but seeing that come up in the disassembly spooked me so I decided to
stop exploring further until I needed to support packing/unpacking in detail.

Added code object flags so we can mark a code object as taking vargs. It would
also be used for kwargs and generators. It is set up to set for kwargs too but
will raise an exception if you try to use the feature.

Code visitor now flags code objects that use vargs and kwargs.

We now need to detect this flag and change how we call stuff that takes
variable arguments.

We can now pack variable arguments (vargs) in a function call