All tests passing after having to contain exceptions
Exceptions would cause the run context not to return. I think this is why it
was an out variable. That's no good now.

I want to see if there's something I can do to have exception escapes come
out when they're not expected because they'll get tucked away in the context
now.