Added support for default arguments, keyword arguments, and variable arguments
Default arguments are declared in the function definitions:

def foo(a, b=1):
  print("b is a default argument")

Keyword overrides let you define arguments out-of-order or specifically name
arguments:

foo(a=1, b=2)

Variable arguments are those *args things you see. They can be named anything,
but must have one star. They become a tuple of arbitrary length, and you can
pass any pile of stuff to them:

def sum(*args):
  total = 0
  for arg in args:
    total += arg
  return total

sum(1, 1, 1,) = 3

Conspiciously missing are **kwargs, the dictionary variable arguments. Also missing
are positional-only and keyword-only specifiers (defined with / or * respectively).

PyTuple was further defined and hardened to support variable arguments, but not
much further and requires much more work. An iterator type was added for it, and it
can contain any type of object.

The interpreter now has to run DURING COMPILATION. This is because defaults have
to be interpreted at the point that a function is defined. So the whole compilation
system now has a bunch of async on it. The "async stink" pervaded the entire test
bench, and now a lot of tests as async Tasks that await on the test finishing. This
made it harder to detect errors because a task could exit with an error yet still
have done enough to pass the comparisons. I think I have done enough to default to
test for exceptions that this shouldn't happen any more.

Add the CALL_FUNCTION_KW opcode in order to handle keyword arguments, it takes a
Tuple of names that correspond to the next values on the stack. Together, they form
the name:value pairs of keyword arguments.

Added code object flags so we can mark a code object as taking vargs. It would
also be used for kwargs and generators. It is set up to set for kwargs too but
will raise an exception if you try to use the feature.

Code visitor now flags code objects that use vargs and kwargs.

We now need to detect this flag and change how we call stuff that takes
variable arguments.

Code generation can recognize defaults and variable arguments and generate the
appropriate code for them.

IScheduler took on more of API so ByteCodeCompiler could take an IScheduler

It's what I was getting at some point in the chain and I decided I'd just
extract what I need from the implementation into the abstraction's signature
because it's been vetted as being necessary for the interface at this
point.

Created ArgParamMatcher to manager Python argument-parameter binding

ArgParamMatcher takes the raw arguments and a CodeObject. It then produces the actual
assignments for the code object to use. This is a decoupled helper that we were able
to test independently. It is not yet integrated into the interpreter code. This
commit just adds it as dark code with test coverage.

It resolves all kinds of combinations of positional, keyword, default, and variable
arguments.

Some things missing:
1. **kwargs
2. positional-only arguments
3. keyword-only arguments

These are noted in the TODO.

I am pondering a state machine reimplementation in the future if it gets too
ugly.

A common helper runs most of CALL_FUNCTION_KW and CALL_FUNCTION since they both
mostly do the same thing except for keyword argument management in CALL_FUNCTION_KW.

# Conflicts:
#	CloacaInterpreter/Interpreter.cs
#	CloacaInterpreter/TODO.md
#	CloacaInterpreter/Visitors.cs
#	CloacaTests/CodeTests.cs
#	LanguageImplementation/CodeObject.cs
#	LanguageImplementation/DataTypes/PyTuple.cs
#	collections.md