We can now parse keyword-only arguments after variable arguments
Stuff like this:

def foo(*args, b=1, c=2):
  pass

The heapq import would choke there, but now chokes on what I think is the list
comprehension. That's going to be a totally different problem.

The ArgParamMatcher has been rewritten to use a state machine. Having each of the
operations in its own state makes the code larger but definitely much easier to
understand. It was the only reasonable way I could add this support. I think it also
makes it clearer when to point out errors.

The visitor logic in the parser wasn't properly transforming negative number constants
in the default fields for function calls. It might have been doing it elsewhere too!
For now, it only manages integers. Next is to manage floats.