Implemented len()
We thought we already have this! The support.md even says as much! Scandal!

I got bit by this while working on double list comprehensions and used len() in a
unit test. It wasn't resolved and I found it wasn't described anywhere. Whoops!

It is implemented with the __len__ dunder in the base types. For .NET types, we
look for a Count property and fall back to a Length property.