Added ImportError as an appropriate, detecting parent class
This is part of the heapq import work. There are some ImportErrors that were
at the end that screwed us up since the interpreter didn't know what
ImportError even was.

* Added ImportError
* Chained ImportError as a parent of ModuleNotFoundError
* Create corresponding classes
* Established rough chain of command between the classes
* Added ImportError and ModuleNotFoundError as builtins

There's a lot of boilerplate for creating these chains and I added a general
TODO to make this much more bearable. Also, I should test this in some way.

The heapq test next needs an awareness of __name__.